Updates for Open Web UI
